product/requirements.txt
Flask
python_dotenv
supabase